Fix compilation error

+++ src/sprite/sprite.cpp
@@ -162,7 +162,7 @@ Sprite::draw(Canvas& canvas, const Vector& pos, int layer,
   else
   {
     canvas.draw_surface(m_action->surfaces[m_frameidx],
-                        pos - Vector(m_action->x_offset, m_action->surfaces[m_frameidx]->get_height() - m_action->y_offset - m_action->hitbox_h),
+                        pos - Vector(m_action->x_offset, static_cast<float>(m_action->surfaces[m_frameidx]->get_height()) - m_action->y_offset - m_action->hitbox_h),
                         m_angle,
                         m_color,
                         m_blend,
